fix(fetcher): Fix circular import issue
