+++ duties/fetcher/identifier/core.py
@@ -5,14 +5,11 @@ from logging import getLogger
 from multiprocessing.shared_memory import SharedMemory
 from pickle import dumps, loads
 from sys import exit as sys_exit
-from typing import Any, Dict, List
+from typing import List
 
-from constants import endpoints, json, logging, program
+from constants import logging, program
 from eth_typing import BLSPubkey
 from fetcher.data_types import ValidatorData, ValidatorIdentifier
-from fetcher.identifier.filter import log_inactive_and_duplicated_validators
-from protocol import ethereum
-from protocol.request import CalldataType, send_beacon_api_request
 
 __LOGGER = getLogger()
 
@@ -197,106 +194,3 @@ def __is_valid_pubkey(pubkey: str, is_logged: bool) -> bool:
             __LOGGER.error(logging.PUBKEY_IS_NOT_HEXADECIMAL_MESSAGE, pubkey, error)
         return False
     return True
-
-
-async def fetch_active_validator_identifiers(
-    provided_raw_validator_identifiers: dict[str, ValidatorIdentifier]
-) -> dict[str, ValidatorIdentifier]:
-    """Fetch active validators based on on-chain status
-
-    Args:
-        provided_raw_validator_identifiers (dict[str, ValidatorIdentifier]): Provided validator
-        identifiers by the user
-    Returns:
-        dict[str, ValidatorIdentifier]: Active validator identifiers
-    """
-    if (
-        len(provided_raw_validator_identifiers)
-        > program.THRESHOLD_TO_INFORM_USER_FOR_WAITING_PERIOD
-    ):
-        __LOGGER.info(
-            logging.HIGHER_PROCESSING_TIME_INFO_MESSAGE,
-            len(provided_raw_validator_identifiers),
-        )
-    provided_validators = [
-        get_validator_index_or_pubkey(None, validator)
-        for validator in provided_raw_validator_identifiers.values()
-    ]
-    validator_infos = await send_beacon_api_request(
-        endpoint=endpoints.VALIDATOR_STATUS_ENDPOINT,
-        calldata_type=CalldataType.PARAMETERS,
-        provided_validators=provided_validators,
-    )
-    provided_active_validator_identifiers = (
-        __create_complete_active_validator_identifiers(
-            validator_infos, provided_validators, provided_raw_validator_identifiers
-        )
-    )
-    return provided_active_validator_identifiers
-
-
-def __create_complete_active_validator_identifiers(
-    fetched_validator_infos: List[Any],
-    provided_validators: List[str],
-    raw_validator_identifiers: dict[str, ValidatorIdentifier],
-) -> Dict[str, ValidatorIdentifier]:
-    """Create complete validator identifiers (index, pubkey, alias) and filters
-    for inactive ones and duplicates
-
-    Args:
-        fetched_validator_infos (List[Any]): Fetched validator infos from the beacon chain
-        provided_validators (List[str]): Provided validators by the user
-
-    Returns:
-        Dict[str, ValidatorIdentifier]: Complete validator identifiers
-        filtered for inactive ones and duplicates
-    """
-    complete_validator_identifiers: Dict[str, ValidatorIdentifier] = {}
-    for validator_info in fetched_validator_infos:
-        raw_identifier = __get_raw_validator_identifier(
-            validator_info, raw_validator_identifiers
-        )
-        if (
-            raw_identifier
-            and validator_info[json.RESPONSE_JSON_STATUS_FIELD_NAME]
-            in ethereum.ACTIVE_VALIDATOR_STATUS
-        ):
-            raw_identifier.index = validator_info[json.RESPONSE_JSON_INDEX_FIELD_NAME]
-            raw_identifier.validator.pubkey = validator_info[
-                json.RESPONSE_JSON_VALIDATOR_FIELD_NAME
-            ][json.RESPONSE_JSON_PUBKEY_FIELD_NAME]
-            complete_validator_identifiers[raw_identifier.index] = raw_identifier
-    log_inactive_and_duplicated_validators(
-        provided_validators, complete_validator_identifiers
-    )
-    return complete_validator_identifiers
-
-
-def __get_raw_validator_identifier(
-    validator_info: Any,
-    raw_validator_identifiers: dict[str, ValidatorIdentifier],
-) -> ValidatorIdentifier | None:
-    """Get raw validator identifier as provided by the user based on the
-    fetched validator infos from the beacon chain
-
-    Args:
-        validator_info (Any): Validator infos from the beacon chain
-
-    Returns:
-        ValidatorIdentifier | None: Raw validator identifier
-    """
-    identifier_index = raw_validator_identifiers.get(
-        validator_info[json.RESPONSE_JSON_INDEX_FIELD_NAME]
-    )
-    identifier_pubkey = raw_validator_identifiers.get(
-        validator_info[json.RESPONSE_JSON_VALIDATOR_FIELD_NAME][
-            json.RESPONSE_JSON_PUBKEY_FIELD_NAME
-        ]
-    )
-    if identifier_index and identifier_pubkey:
-        if identifier_index.alias:
-            return identifier_index
-        return identifier_pubkey
-    if identifier_index:
-        return identifier_index
-    return identifier_pubkey
